+++ bidon/src/main/java/org/bidon/sdk/stats/models/BidStat.kt
@@ -7,6 +7,7 @@ import org.bidon.sdk.adapter.DemandId
 data class BidStat(
     val auctionId: String?,
     val roundId: String?,
+    val roundIndex: Int?,
     val demandId: DemandId,
     val roundStatus: RoundStatus?,