+++ bidon/src/main/java/org/bidon/sdk/auction/usecases/impl/AuctionStatImpl.kt
@@ -94,7 +94,7 @@ internal class AuctionStatImpl(
                         ).code
                     )
                 },
-                bidding = roundStat.bidding?.copy(
+                bidding = roundStat.bidding?.takeIf { it.bids.isNotEmpty() }?.copy(
                     bids = roundStat.bidding.bids.map { bid ->
                         bid.copy(
                             roundStatusCode = RoundStatus.values().first {
@@ -382,7 +382,8 @@ internal class AuctionStatImpl(
             adUnitId = stat?.adUnitId.takeIf { isSucceed },
             auctionStartTs = auctionStartTs,
             auctionFinishTs = auctionFinishTs,
-            roundId = stat?.roundId
+            roundId = stat?.roundId,
+            bidType = stat?.bidType?.code,
         )
     }
 }