+++ bidon/src/main/java/org/bidon/sdk/auction/usecases/impl/ExecuteRoundUseCaseImpl.kt
@@ -29,6 +29,7 @@ import org.bidon.sdk.logs.logging.impl.logError
 import org.bidon.sdk.logs.logging.impl.logInfo
 import org.bidon.sdk.regulation.Regulation
 import org.bidon.sdk.stats.StatisticsCollector
+import org.bidon.sdk.stats.models.BidType
 
 internal class ExecuteRoundUseCaseImpl(
     private val adaptersSource: AdaptersSource,
@@ -69,7 +70,8 @@ internal class ExecuteRoundUseCaseImpl(
                         auctionResponse = auctionResponse,
                         demandAd = demandAd,
                         round = round,
-                        roundIndex = roundIndex
+                        roundIndex = roundIndex,
+                        bidType = BidType.RTB
                     )
                 }
                 .filterIsInstance<Mode.Bidding>()
@@ -111,7 +113,8 @@ internal class ExecuteRoundUseCaseImpl(
                         auctionResponse = auctionResponse,
                         demandAd = demandAd,
                         round = round,
-                        roundIndex = roundIndex
+                        roundIndex = roundIndex,
+                        bidType = BidType.CPM
                     )
                 }
                 .filterIsInstance<Mode.Network>()
@@ -174,13 +177,15 @@ internal class ExecuteRoundUseCaseImpl(
         auctionResponse: AuctionResponse,
         demandAd: DemandAd,
         round: RoundRequest,
-        roundIndex: Int
+        roundIndex: Int,
+        bidType: BidType
     ) {
         adSource.addRoundInfo(
             auctionId = auctionResponse.auctionId,
             roundId = round.id,
             demandAd = demandAd,
-            roundIndex = roundIndex
+            roundIndex = roundIndex,
+            bidType = bidType
         )
         adSource.setStatisticAdType(adTypeParam.asStatisticAdType())
         adSource.addAuctionConfigurationId(auctionResponse.auctionConfigurationId ?: 0)