+++ bidon/src/main/java/org/bidon/sdk/stats/models/BidStat.kt
@@ -15,4 +15,6 @@ data class BidStat(
     val fillStartTs: Long?,
     val fillFinishTs: Long?,
     val adUnitId: String?,
+
+    val bidType: BidType?,
 )