+++ bidon/src/main/java/org/bidon/sdk/stats/models/ResultBody.kt
@@ -13,6 +13,8 @@ internal data class ResultBody(
     val roundId: String?,
     @field:JsonName("winner_id")
     val demandId: String?,
+    @field:JsonName("bid_type")
+    val bidType: String?,
     @field:JsonName("ecpm")
     val ecpm: Double?,
     @field:JsonName("ad_unit_id")