+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Release 0.4.0 (2023.09.13)
+## Bugfixes:
+[BDN-391](https://appodeal.atlassian.net/browse/BDN-391) Sync Adapter version between Appodeal vs Bidon
+
 # Release 0.4.0 (2023.09.08)
 ## New Features:
 [BDN-377](https://appodeal.atlassian.net/browse/BDN-377) Add meditation_service to Meta Audience Adapter